# docs/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: Inventory Management System API
  description: |-
    Backend contract for the Inventory Management System frontend clients.

    Authentication: obtain a JWT via POST /auth/login or POST /auth/register,
    then send `Authorization: Bearer <accessToken>` on every protected call.
    Refresh tokens are exchanged at POST /auth/refresh.

    Real-time updates are delivered over STOMP over WebSocket at /api/ws
    (see docs/api-contract/asyncapi-ws.md) - they are not part of this
    OpenAPI document.
  version: 0.1.0
servers:
- url: http://localhost
  description: Generated server url
security:
- bearerAuth: []
...
paths:
...
  /api/orders/{id}/ship:
    post:
      tags:
      - Orders
      summary: Ship a paid order
      description: Transitions a paid order to the shipped state.
      operationId: shipOrder
      parameters:
      - name: id
        in: path
        description: Unique order identifier
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Order shipped
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderResponse"
  /orders/{id}/ship:
    post:
      tags:
      - Orders
      summary: Ship a paid order
      description: Transitions a paid order to the shipped state.
      operationId: shipOrder_1
      parameters:
      - name: id
        in: path
        description: Unique order identifier
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Order shipped
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderResponse"
  /api/orders/{id}/payment:
    post:
      tags:
      - Orders
      summary: Process payment for an order
      description: Transitions a pending order to a paid state.
      operationId: processPayment
      parameters:
      - name: id
        in: path
        description: Unique order identifier
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Payment processed
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderResponse"
  /orders/{id}/payment:
    post:
      tags:
      - Orders
      summary: Process payment for an order
      description: Transitions a pending order to a paid state.
      operationId: processPayment_1
      parameters:
      - name: id
        in: path
        description: Unique order identifier
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Payment processed
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderResponse"
  /orders/{id}/notify-user/{username}:
    post:
      tags:
      - Orders
      operationId: notifyUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
  /api/orders/{id}/notify-user/{username}:
    post:
      tags:
      - Orders
      operationId: notifyUser_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
  /api/orders/{id}/notify-test:
    post:
      tags:
      - Orders
      operationId: notifyTest
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
  /orders/{id}/notify-test:
    post:
      tags:
      - Orders
      operationId: notifyTest_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
...
  /orders/{id}/cancel:
    post:
      tags:
      - Orders
      summary: Cancel an order
      description: Cancels an order when allowed by the domain lifecycle.
      operationId: cancelOrder
      parameters:
      - name: id
        in: path
        description: Unique order identifier
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Order cancelled
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderResponse"
  /api/orders/{id}/cancel:
    post:
      tags:
      - Orders
      summary: Cancel an order
      description: Cancels an order when allowed by the domain lifecycle.
      operationId: cancelOrder_1
      parameters:
      - name: id
        in: path
        description: Unique order identifier
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Order cancelled
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderResponse"
  /orders:
    get:
      tags:
      - Orders
      summary: List orders
      description: "Lists orders for a customer, optionally filtered by status. Non-admin\
        \ callers are always scoped to their own orders."
      operationId: listOrders
      parameters:
      - name: customerId
        in: query
        description: Optional customer identifier (admin only)
        required: false
        schema:
          type: string
      - name: status
        in: query
        description: Optional status filter
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Orders listed
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderResponse"
    post:
      tags:
      - Orders
      summary: Create a new order
      description: Creates a pending order from a valid request payload.
      operationId: createOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
        required: true
      responses:
        "201":
          description: Order created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          description: Invalid request body
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderResponse"
  /api/orders:
    get:
      tags:
      - Orders
      summary: List orders
      description: "Lists orders for a customer, optionally filtered by status. Non-admin\
        \ callers are always scoped to their own orders."
      operationId: listOrders_1
      parameters:
      - name: customerId
        in: query
        description: Optional customer identifier (admin only)
        required: false
        schema:
          type: string
      - name: status
        in: query
        description: Optional status filter
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Orders listed
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderResponse"
    post:
      tags:
      - Orders
      summary: Create a new order
      description: Creates a pending order from a valid request payload.
      operationId: createOrder_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
        required: true
      responses:
        "201":
          description: Order created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          description: Invalid request body
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderResponse"
...
  /orders/{id}:
    get:
      tags:
      - Orders
      summary: Get one order
      description: Fetches an order by its identifier.
      operationId: getOrder
      parameters:
      - name: id
        in: path
        description: Unique order identifier
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Order found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderResponse"
  /api/orders/{id}:
    get:
      tags:
      - Orders
      summary: Get one order
      description: Fetches an order by its identifier.
      operationId: getOrder_1
      parameters:
      - name: id
        in: path
        description: Unique order identifier
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Order found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderResponse"
  /orders/status/{orderId}:
    get:
      tags:
      - Orders
      summary: Get order status
      description: Returns the current status of an order.
      operationId: getOrderStatus
      parameters:
      - name: orderId
        in: path
        description: Unique order identifier
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Order status returned
          content:
            '*/*':
              schema:
                type: object
                additionalProperties: {}
  /api/orders/status/{orderId}:
    get:
      tags:
      - Orders
      summary: Get order status
      description: Returns the current status of an order.
      operationId: getOrderStatus_1
      parameters:
      - name: orderId
        in: path
        description: Unique order identifier
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Order status returned
          content:
            '*/*':
              schema:
                type: object
                additionalProperties: {}
...
  /api/fulfillment/orders:
    get:
      tags:
      - Orders
      summary: Fulfillment queue
      description: Lists PAID/SHIPPED/DELIVERED orders across all customers for warehouse
        processing.
      operationId: fulfillmentQueue
      parameters:
      - name: status
        in: query
        description: "Optional status filter (PAID, SHIPPED, DELIVERED)"
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Fulfillment queue returned
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderResponse"
  /fulfillment/orders:
    get:
      tags:
      - Orders
      summary: Fulfillment queue
      description: Lists PAID/SHIPPED/DELIVERED orders across all customers for warehouse
        processing.
      operationId: fulfillmentQueue_1
      parameters:
      - name: status
        in: query
        description: "Optional status filter (PAID, SHIPPED, DELIVERED)"
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Fulfillment queue returned
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderResponse"
...
  /api/admin/audit/orders/{id}:
    get:
      tags:
      - Audit
      summary: Order change history
      description: Returns the Envers revision history for an order.
      operationId: orderHistory
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Audit history returned
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditEntryResponse"
...
components:
  schemas:
...
    OrderResponse:
      type: object
      description: Order payload returned by the REST API
      properties:
        id:
          type: string
          description: Order identifier
          example: ord-123
        customerId:
          type: string
          description: Customer identifier
          example: customer-001
        status:
          type: string
          description: Order lifecycle status
          example: PENDING
        totalAmount:
          type: number
          description: Order total amount
          example: 25.0
        currency:
          type: string
          description: Currency code
          example: USD
        createdAt:
          type: string
          format: date-time
          description: Order creation timestamp
        reservedUntil:
          type: string
          format: date-time
          description: Reservation expiry timestamp
        reservationSecondsRemaining:
          type: integer
          format: int64
          description: Server-computed seconds left before the reservation expires
            (0 when not pending)
    CreateOrderItemRequest:
      type: object
      description: Order item payload for a product reservation
      properties:
        productId:
          type: string
          description: Identifier of the product to reserve
          example: SKU-001
          minLength: 1
        quantity:
          type: integer
          format: int32
          description: Requested item quantity
          example: 2
          minimum: 1
      required:
      - productId
    CreateOrderRequest:
      type: object
      description: Payload used to create a new order
      properties:
        customerId:
          type: string
          description: Customer identifier
          example: customer-001
          minLength: 1
        items:
          type: array
          description: Order items to reserve
          items:
            $ref: "#/components/schemas/CreateOrderItemRequest"
          minItems: 1
      required:
      - customerId
      - items
...
    AuditEntryResponse:
      type: object
      properties:
        revision:
          type: integer
          format: int64
        timestamp:
          type: integer
          format: int64
        author:
          type: string
        revisionType:
          type: string
        snapshot:
          type: object
          additionalProperties: {}
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT